server/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
azure-ai-formrecognizer
azure-ai-textanalytics
openai
